example/requirements.txt
aioredis==1.1.0
aniso8601==3.0.2
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
channels==2.1.3
channels-redis==2.3.0
constantly==15.1.0
daphne==2.2.2
Django==2.1.2
django-channels-graphql-ws==0.1.15
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
msgpack==0.5.6
msgpack-python==0.5.6
namedlist==1.7
promise==2.2.1
PyHamcrest==1.9.0
pytz==2018.5
redis==2.10.6
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
Twisted==18.7.0
txaio==18.8.1
zope.interface==4.5.0
